requirements.txt
acres==0.3.0
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.1.8
configobj==5.0.9
configparser==7.2.0
distro==1.9.0
etelemetry==0.3.1
exceptiongroup==1.3.0
fastapi==0.115.12
filelock==3.18.0
fitz==0.0.1.dev2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
isodate==0.6.1
jiter==0.10.0
looseversion==1.3.0
lxml==5.4.0
networkx==3.2.1
nibabel==5.3.2
nipype==1.10.0
numpy==2.0.2
openai==1.79.0
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.2.3
pathlib==1.0.1
pdf2image==1.17.0
pillow==11.2.1
prov==2.0.1
psutil==7.0.0
puremagic==1.29
pydantic==2.11.4
pydantic_core==2.33.2
pydot==4.0.0
PyMuPDF==1.25.5
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
pyxnat==1.6.3
rdflib==6.3.2
requests==2.32.3
scipy==1.13.1
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
traits==7.0.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
XlsxWriter==3.2.3
zipp==3.21.0